refactor: migrate learn views into sites/brc-analytics and learn icons into shared (#1557)
