+++ packages/shared/components/CustomIcon/components/LiveHelpIcon/liveHelpIcon.tsx
@@ -1,9 +1,11 @@
 import { PALETTE } from "@databiosphere/findable-ui/lib/styles/common/constants/palette";
-import { SvgIcon, type SvgIconProps } from "@mui/material";
-import { type JSX } from "react";
+import { SVG_ICON_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/svgIcon";
+import type { SvgIconProps } from "@mui/material";
+import { SvgIcon } from "@mui/material";
+import { JSX } from "react";
 
 export const LiveHelpIcon = ({
-  fontSize = "large",
+  fontSize = SVG_ICON_PROPS.FONT_SIZE.LARGE,
   viewBox = "0 0 48 48",
   ...props
 }: SvgIconProps): JSX.Element => {
