+++ pages/learn/index.tsx
@@ -1,9 +1,9 @@
 import { BRC_PAGE_META } from "@/common/meta/brc/constants";
 import { StyledPagesMain } from "@/components/Layout/components/Main/main.styles";
-import { LearnView } from "@/views/LearnView/learnView";
-import type { PageProps } from "@pages/_app";
-import { type GetStaticProps } from "next";
-import { type JSX } from "react";
+import { LearnView } from "@brc/views/LearnView/learnView";
+import { GetStaticProps } from "next";
+import { JSX } from "react";
+import type { PageProps } from "../_app";
 
 const Page = (): JSX.Element => {
   return <LearnView />;
